feat: Adicionar novos tipos de serviços e refatorar código para melhor legibilidade
